==> season23_24/notebooks/8_fplreview_style_projection_data.ipynb <==
This notebook is used to create a csv file that includes my point projections in a format similar to fplreview.com. This csv file can then be used for making advanced team optimization plans using Sertalp B. Cay's repo "fpl-optimization".

latest_gameweek = 37

import numpy as np
import pandas as pd
from pathlib import Path
import json

# Save team data
Info on your own team; starting point for the optimization.

# get login credentials for fetching team data
file_path = Path('../../login_credentials/fpl_login.json')
login_credentials = pd.read_json(file_path, typ='series')

# utility function for fetching team data
from src.utils import fetch_my_team

# fetch my team data
my_team = fetch_my_team(login_credentials.user_name, login_credentials.password, login_credentials.team_id)

# save team data to fpl-optimization repo
file_path = Path('../../../repos/fpl-optimization/data/team.json')
with open(file_path, 'w') as f:
    json.dump(my_team, f)

# Edit projections into fplreview format 
### (to be used with fpl-optimization repo solver)

filepath = Path(f'../data/predictions/gameweek{latest_gameweek}.csv')
projections = pd.read_csv(filepath, index_col=0)
display(projections.head())
display(projections.shape)

filepath = Path('../data/fpl_df.csv')
fpl_df = pd.read_csv(filepath, index_col=0, low_memory=False)
fpl_df = fpl_df[fpl_df.season=='23-24']
display(fpl_df.head())
display(fpl_df.shape)

df = fpl_df.groupby('name').last().reset_index()[['id', 'name','points_per_game', 'total_points',]]
#df['id'] = df.id.astype(int)
df['games_played'] = np.round(np.where(df['points_per_game']!=0, df['total_points'] / df['points_per_game'], 0),0)
display(df.head())
display(df.shape)

# drop duplicate players (some players get new spelling for their name during the season causing duplicates)
duplicate_ids = df.loc[df.id.duplicated(), 'id'].unique()
for id in duplicate_ids:
    ix = df.loc[df.id==id, 'games_played'].idxmin()
    df = df.drop(ix)
display(df.head())
display(df.shape)

# drop unneccesary columns
df = df.drop(['points_per_game', 'total_points','games_played'], axis=1)
display(df.head())
display(df.shape)

# merge id info to projections
projections = projections.merge(df, on='name', how='left')
#projections.head()

# add xMins variable that is needed later
projections['xMins'] = 90

projections[['id', 'expected_points', 'xMins', 'gameweek']]

projections_pivot = (
    projections
    .pivot_table(
    columns=['gameweek',],
    index='id', 
    values=['expected_points','xMins'], 
    aggfunc='sum'
    )
)

projections_pivot.head()

new_cols = []
for col in projections_pivot.columns:
    if col[0] == 'expected_points':
        new_col = str(col[1]) + '_Pts'
        new_cols.append(new_col)
    elif col[0] == 'xMins':
        new_col = str(col[1]) + '_xMins'
        new_cols.append(new_col)

projections_pivot.columns = new_cols
projections_pivot = projections_pivot.reset_index()
projections_pivot = projections_pivot.rename(columns={'id':'ID'})

projections_pivot.head()

filepath = Path('../../../repos/fpl-optimization/data/fplreview.csv')
projections_pivot.to_csv(filepath)

### Option to make manual changes, e.g., to xPts (based on injuries)

### De Bruyne

projections.loc[projections['name'].str.contains('Bruy'), 'id'].unique()

player_id = 349
display(projections.loc[projections.id==player_id, 'name'].unique())
display(projections_pivot[projections_pivot.ID==player_id])

#player_id = 353
gameweeks = np.arange(latest_gameweek+1,39)
weights = [0.75]

for i in range(len(gameweeks)):
    projections_pivot.loc[projections_pivot.ID==player_id, f'{gameweeks[i]}_Pts'] = \
        weights[i]*projections_pivot.loc[projections_pivot.ID==player_id, f'{gameweeks[i]}_Pts']

display(projections_pivot[projections_pivot.ID==player_id])

### Dubravka

projections.loc[projections['name'].str.contains('Dubrav'), 'id'].unique()

player_id = 409
display(projections.loc[projections.id==player_id, 'name'].unique())
display(projections_pivot[projections_pivot.ID==player_id])

gameweeks = [38]
weights = [0.5]

for i in range(len(gameweeks)):
    projections_pivot.loc[projections_pivot.ID==player_id, f'{gameweeks[i]}_Pts'] = \
        weights[i]*projections_pivot.loc[projections_pivot.ID==player_id, f'{gameweeks[i]}_Pts']

display(projections_pivot[projections_pivot.ID==player_id])

### Turner

projections.loc[projections['name'].str.contains('Turner'), 'id'].unique()

player_id = 28
projections_pivot[projections_pivot.ID==player_id]

gameweeks = np.arange(latest_gameweek+1,39)
weights = np.zeros(len(gameweeks))

for i in range(len(gameweeks)):
    projections_pivot.loc[projections_pivot.ID==player_id, f'{gameweeks[i]}_Pts'] = \
        weights[i]*projections_pivot.loc[projections_pivot.ID==player_id, f'{gameweeks[i]}_Pts']

display(projections_pivot[projections_pivot.ID==player_id])

### Ederson

projections.loc[projections['name'].str.contains('Eders'), 'id'].unique()

player_id = 352
projections_pivot[projections_pivot.ID==player_id]

gameweeks = np.arange(latest_gameweek+1,39)
weights = np.zeros(len(gameweeks))

for i in range(len(gameweeks)):
    projections_pivot.loc[projections_pivot.ID==player_id, f'{gameweeks[i]}_Pts'] = \
        weights[i]*projections_pivot.loc[projections_pivot.ID==player_id, f'{gameweeks[i]}_Pts']

display(projections_pivot[projections_pivot.ID==player_id])

# save data
filepath = Path('../../../repos/fpl-optimization/data/fplreview.csv')
projections_pivot.to_csv(filepath)

# check ids
projections.loc[projections['name'].str.contains('Tripp'), ['id','name']]

# banned players from solver 
projections.loc[projections.id.isin([77,343,31,301,10,230,590,546,145,424,368,216,148,152,203,195,197]), 'name'].unique()
